requirements.txt
Pillow==8.4.0
selenium==4.0.0
webdriver_manager==3.5.1
